requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
SQLAlchemy==1.4.41
python-dotenv==1.0.0
PyMySQL==1.1.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
email-validator==2.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
cryptography==41.0.4  # Required for PyMySQL